fix(FEC-14472): part of the caption isn't shown on playback on VOD entry (#227)
### Description of the Changes

**the issue:**
part of the captions are not loaded when the following conditions are met:
- starting the playback with disabled captions
- captions are webvtt
- the captions are long (more than 5 minutes) and there are webvtt segments
- playing hls
- not native captions

**root cause:**
- first webvtt chunk is loaded
- since captions are started off - we are hiding the captions by setting `hls.subtitleTrack=-1`
- this is causing hls to disable the captions and stop the process of parsing the next chunk
- hls is not loading captions retroactively, hence, captions are "lost"

**solution:**
since the use-case is for text tracks that are being rendered by kaltura player, we are controlling the visibility of the captions. we already handle resetting the view - we don't need to set `hls.subtitleTrack=-1` which is causing issues.

#### Resolves FEC-14472